pyproject.toml
[project]
name = "apelr"
version = "0.1.0"
description = "APEL-R prototype: asynchronous planner-executor language model"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "torch",
  "datasets",
  "numpy",
  "pyyaml",
  "tqdm",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apelr"]

